SYCL: fix pointer arithmetic for very large matrices (see 3e51680 and 2c67704 for CUDA commits)

+++ magmablas_sycl/ztranspose.dp.cpp
@@ -57,8 +57,8 @@ ztranspose_device(
     int iby = item_ct1.get_group(1) * NB;
     int i, j;
 
-    A  += ibx + tx + (iby + ty)*lda;
-    AT += iby + tx + (ibx + ty)*ldat;
+    A  += ibx + tx + (size_t)(iby + ty)*(size_t)lda;
+    AT += iby + tx + (size_t)(ibx + ty)*(size_t)ldat;
 
     #pragma unroll
     for( int tile=0; tile < NB/NX; ++tile ) {
@@ -103,7 +103,7 @@ ztranspose_device(
 
         // move to next subtile
         A  += NX;
-        AT += NX*ldat;
+        AT += (size_t)NX*(size_t)ldat;
     }
 }
